# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ecomotif API
  description: |
    Spesifikasi OpenAPI diselaraskan dengan `docs/API.md`.
    Server `url: /api` — path di bawah ini **tanpa** prefix `/api` kedua.
    **JWT** (`auth:api`): header `Authorization: Bearer <token>`.
    **Sanctum** (`/v1/*`): header `Authorization: Bearer <sanctum_token>`.
  version: "1.2.0"
  externalDocs:
    description: Parameter lengkap & catatan implementasi
    url: ./API.md
servers:
  - url: /api
    description: Relatif ke origin aplikasi (APP_URL)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CarBasicCreate:
      type: object
      required: [brand_id, title, slug, description, condition, vehicle_type, purpose, regular_price]
      properties:
        brand_id: { type: integer }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        condition: { type: string, enum: [used, new] }
        vehicle_type: { type: string, enum: [car, motorcycle] }
        purpose: { type: string, enum: [Rent, Sale] }
        regular_price: { type: number }
        offer_price: { type: number }
        seo_title: { type: string }
        seo_description: { type: string }
...
paths:
...
  /user/car:
    get:
      tags: [Car]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      tags: [Car]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarBasicCreate"
      responses:
        "200": { description: OK }
...
  /user/car/{car}:
    get:
      tags: [Car]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: car
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK }
    put:
      tags: [Car]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: car
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CarBasicCreate"
                - type: object
                  properties:
                    lang_code: { type: string }
                    translate_id: { type: integer }
      responses:
        "200": { description: OK }
    delete:
      tags: [Car]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: car
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK }
...